requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
fastai>=2.3.1
ipywidgets
voila
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
numpy==1.21.5
Pillow==9.0.1
pyTelegramBotAPI==4.4.0
requests==2.27.1
typing-extensions==4.0.1
urllib3==1.26.8
